pyproject.toml
[project]
name = "zl-scraper"
version = "0.1.0"
description = "ZnanyLekarz clinic scraping pipeline"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "typer[all]>=0.12",
    "tenacity>=8.2",
    "python-dotenv>=1.0",
]

[project.scripts]
zl-scraper = "zl_scraper.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/zl_scraper"]
